float attributes are now working.

+++ graf2d/primitives/v7/inc/ROOT/TText.hxx
@@ -41,9 +41,9 @@ public:
  */
 
 class DrawingOpts: public TDrawingOptsBase {
-   TDrawingAttr<TColor> fTextColor{*this, "Text.Color", TColor::kBlack};   ///< The text color.
-   TDrawingAttr<int>    fTextSize{*this, "Text.Size", 10};                 ///< The text size
-   TDrawingAttr<int>    fTextAngle{*this, "Text.Angle", 0};                ///< The text angle
+   TDrawingAttr<TColor> fTextColor{*this, "Text.Color", TColor::kBlack}; ///< The text color.
+   TDrawingAttr<float>  fTextSize{*this, "Text.Size", 10.};              ///< The text size
+   TDrawingAttr<float>  fTextAngle{*this, "Text.Angle", 0.};             ///< The text angle
 
 public:
    /// The color of the text.
@@ -52,14 +52,14 @@ public:
    const TColor &GetTextColor() const   { return fTextColor.Get(); }
 
    /// The text size.
-   void SetTextSize(int size) { fTextSize = size; }
-   TDrawingAttr<int> &GetTextSize() { return fTextSize; }
-   int GetTextSize() const { return (int)fTextSize; }
+   void SetTextSize(float size) { fTextSize = size; }
+   TDrawingAttr<float> &GetTextSize() { return fTextSize; }
+   float GetTextSize() const { return (float)fTextSize; }
 
    /// The text angle in degrees.
-   void SetTextAngle(int angle) { fTextAngle = angle; }
-   TDrawingAttr<int> &GetTextAngle() { return fTextAngle; }
-   int GetTextAngle() const { return (int)fTextAngle; }
+   void SetTextAngle(float angle) { fTextAngle = angle; }
+   TDrawingAttr<float> &GetTextAngle() { return fTextAngle; }
+   float GetTextAngle() const { return (float)fTextAngle; }
 };
